Change output of  RunTable filtering methods + add new methods
Changed the RunTable filtering methods to return new RunTable instances instead
of pandas series. This was done to ensure that the filtered datasets resulting
from the filtering methods would still have access to the RunTable filtering
methods so that the user could further-refine the dataset if needed.

This change requires that the RunTable.get_accession_numbers() method be added,
since the filtering methods no longer return lists of accession numbers.

This change also makes it logically sensible to add the
RunTable.random_sample_subset() instance method to the class, whereas previously
it would have needed to be added as a @staticmethod since it would have been
operating on lists of accession numbers that were not instances of RunTable.