Enable relative path usage in file handling via BioPype.workspaces
Needed a way to reference the paths to other files in the BioPype module.
Created package-wide variables that the user can specify to store
absolute path names to the BioPype and SRA Toolkit working directory
and Workspace, respectively. The absolute paths can be joined with
relative paths to navigate and manipulate files throughout the BioPype directory.