Modify DirPathsHelper class, BioPype __init__.py, and downloadsra.py
Added greeting() method to DirPathsHelper class. This method is called in the
__init__.py file for the BioPype package and handles the verbose checking of
the BioPype and SRA Toolkit working directories/workspaces. Separating this functionality
from the code that actually defines the _BIODIR and _SRADIR variables means that
other files can import the DirPathsHelper class and access the two path variables
without triggering the verbose output (because having those prompts pop up
repeatedly would be irritating).

The BIODIR and SRADIR attributes were renamed to _BIODIR and _SRADIR to indicate
private names.

downloadsra.py now imports DirPathsHelper and accesses the paths to the working
directories through the class attributes. This method should serve to port the
paths to other files as well. With this method, separate files can share the same
relative paths without needing to specify those paths with global variables.